Check for pageant on Windows
The `simplessh` package uses pageant for its SSH agent on Windows.
Update the `util.HashSSHAgent` function to detect pageant if running on Windows.
Note that the package `github.com/davidmz/go-pageant` is now a direct dependency,
and the version used is the same version required by the `simplessh` package.

Fixes #24